refactor(spf): rename composition/engine.ts → create-composition.ts
The engine.ts name was a holdover from when createComposition was still
called createPlaybackEngine (renamed in d9cac1b7). Now that the file
lives under core/composition/ alongside the other framework primitive
factories, match the established naming pattern:

- create-machine-actor.ts    → createMachineActor
- create-machine-reactor.ts  → createMachineReactor
- create-transition-actor.ts → createTransitionActor
- create-composition.ts      → createComposition  (was engine.ts)

Rename the three test files alongside — create-composition.test.ts,
create-composition-types.test.ts, create-composition-types.test-d.ts.
Update every importer: src/index.ts, dom/playback-engine/hls-engine.ts,
dom/playback-engine/adapter.ts, internal tests, and the path reference
in internal/design/spf/architecture.md.

No API changes — only the module path moves.
